test: gitea-operator
